Push members down
